Sprint_n2/requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Faker==37.11.0
fakers==0.5.0
h11==0.16.0
idna==3.10
iniconfig==2.3.0
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
py4j==0.10.9.9
Pygments==2.19.2
PySocks==1.7.1
pyspark==4.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
